==> 03 Scripts/2. Python Data Types.ipynb <==
#01. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

print(x)

type(x)

z = 'Brazil'

type(z)

y = 5.678

type(y)

k = False

type(k)

# Working with data types

print(33+33+34)

x = 4.2

y=2.1

print(x/y)

j= 'Hello'

m= 'My name is Vinh'

print(j+m)
