==> 03 Scripts/3. Data import and descriptive analysis.ipynb <==
# Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

df = pd.read_csv(r'/Users/vinhdao/Documents/Instacart Basket Analysis/02 Data/Original Data/orders.csv', index_col = False)

path = '/Users/vinhdao/Documents/Instacart Basket Analysis'

path

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df.head(30)

df.tail()


df.shape

df.info()
