==> 03 Scripts/4. Data Wrangling and Subsetting.ipynb <==
# Define the variables and sections


path = '/Users/vinhdao/Documents/Instacart Basket Analysis'

# Import libraries
import pandas as pd
import numpy as np
import os

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data','products.csv'), index_col = False)

df_ords.drop(columns = ['eval_set'])


df_ords.head(30)

df_ords_2 = df_ords.drop(columns = ['eval_set'])


df_ords_2.head(30)

df_ords['days_since_prior_order'].value_counts(dropna=False)


df_ords['eval_set'].value_counts(dropna=False)


df_ords.head(30)

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data','departments.csv'), index_col = False)

df_dep.head()

# Tranposing df_dep
df_dep.T

# The previous command just change the visual. To change completely, we need to change by assigning to a new variables:
df_dep_t = df_dep.T


df_dep_t.head()

df_dep_t

df_dep_t.reset_index()

new_header = df_dep_t.iloc[0]

new_header

df_dep_t_new = df_dep_t[1:]

df_dep_t.reset_index()

df_dep_t_new.columns = new_header # set the header row as the df header

df_dep_t_new

data_dict = df_dep_t_new.to_dict('index')


data_dict

df_prods.head()

print(data_dict.get('19'))


# Task

#Step 2
df_ords['order_id'] = df_ords['order_id'].astype('str')

# Step 3: changing the name of order_dow column
df_ords.rename(columns={'order_dow': 'order_day_of_week'}, inplace=False)

# Step 4: Find the busiest hour of the way
busiest_hour = df_ords['order_hour_of_day'].value_counts().idxmax()
print(f"The busiest hour for placing orders is: {busiest_hour}")


# Step 5: What "4" means in the dictionary
print(data_dict.get('4'))


# Step 6:
df_breakfast = df_prods[df_prods['department_id'] == 14]


df_breakfast

# Step 7
df_prods_dinner = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]


df_prods_dinner

# Dinner data frame dimension (step 8)
df_prods_dinner.shape

# Step 9
user_1_info = df_ords[df_ords['user_id'] == 1]


user_1_info

# Step 10
user_1_info.describe()


# Step 11
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

# Step 12
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
