==> 03 Scripts/4.8 pr.ipynb <==
# Import libraries

import pandas as pd
import numpy as np
import os

path = '/Users/vinhdao/Documents/Instacart Basket Analysis'
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))

# Create a subset

df = ords_prods_merge[:1000000]

df.shape

df.head(10)

## Grouping Data with pandas

df.groupby('product_name')

# Performing a single aggregation

df.groupby('department_id').agg({'order_number': ['mean']})

Here, the groupby() function is being assigned to the df dataframe. This creates the pandas object for “department_id.” Onto this resulting object, the agg() function is applied. The agg() function will return the mean of the given column, in this case, “order_number.” Once executed, you should see something like this:

# Perform aggregation without using agg function

df.groupby('department_id')['order_number'].mean()

### Note: 

Remember the key difference in syntax between the two methods: when using agg(), put the column you want to aggregate inside the parentheses of the agg() function as an argument. When using mean() (or any other standard aggregation function), simply index the column with square brackets, then follow it with the function you want to use after the dot.

# Performing multiple aggregation

df.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

ords_prods_merge.head(100)

ords_prods_merge =ords_prods_merge.sort_values(by='max_order', ascending=False)


ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'
ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

ords_prods_merge.head(100)
