==> 03 Scripts/5. Data Consistency check.ipynb <==
# Define the variables and sections


path = '/Users/vinhdao/Documents/Instacart Basket Analysis'

# Import libraries
import pandas as pd
import numpy as np
import os

df_ords = pd.read_csv(os.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data','products.csv'), index_col = False)

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# Task

# Find the sum of the missing data
df_prods.isnull().sum()

# Find the NaN value
df_nan = df_prods[df_prods['product_name'].isnull()  == True]

df_nan

df_prods.describe()

df_prods.shape

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]


df_prods_clean

df_prods.dropna(subset = ['product_name'], inplace = True)

df_prods.shape

df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

# Step 2

df_ords.shape

df_ords.describe()

# Interpretation of result

print('High order_number: The maximum value of 100 for order_number might warrant investigation to ensure these values are accurate and not due to data entry errors.')

print('days_since_prior_order:')
print('The presence of 0 could indicate consecutive day orders or might need to be checked for potential data recording issues.')
print('The capping at 30 days should be verified if it aligns with the business logic or data collection constraints.')

for col in df_ords.columns.tolist():
    weird = (df_ords[col].map(type) != df_ords[col].iloc[0].__class__).any()
    if weird:
        print(col)

df_ords.info()

# Step 5: Find the missing value
df_ords.isnull().sum()

# From the previous checkup, we can see that the days_since_prior_order has the large number of missing value

# Step 6: Create subset of the dataframe that contains the nulls
df_ords_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_ords_nan

df_ords.dropna(subset = ['days_since_prior_order'], inplace = True)

df_ords.isnull().sum()

df_ords.shape

df_dups = df_ords[df_ords.duplicated()]

df_dups

There are no duplicate data in this

# Step 9
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_cleaned.csv'))

# Step 10
df_prods.to_csv(os.path.join(path, '02 Data','Prepared Data', 'prods_cleaned.csv'))
