==> 03 Scripts/6.1 Combining Data.ipynb <==
# Merging and Exporting Data

### The script will follow the steps:

1. Import Data

 2. Check Output of Data

 3. Merge dataframes df_ords and df_ords_prior

# Import libraries
import pandas as pd
import numpy as np
import os

### 1. Import Data

path = '/Users/vinhdao/Documents/Instacart Basket Analysis'

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

### 2. Check Output Data

# Check the output: df_ords
df_ords.head()

df_ords.describe()

df_ords.shape

# Check the output
df_prods.head()

df_prods.describe()

df_prods.shape

# Check the output: df_ords_prior
df_ords_prior.head()

df_ords_prior.describe()

df_ords_prior.shape

### 3. Merge dataframes df_ords and df_ords_priors

#Merge df_ords and df_ords_prior
df_merged = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

#Check output of merged dataframe
df_merged_large.head()

df_merged_large['_merge'].value_counts()

#Export data to pkl
df_merged_large.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))
