==> 03 Scripts/6.2 Merging and Exporting Data.ipynb <==
# Merging and Exporting Data

## This script includes the following points:

 Step 3: Import the data

Step 4: Check the Output of Dataframes

Step 5: Merge df_prods with df_ords_combined.

Step 6: Confirm the results of the merge using merge flag

Ste 7: Export the file

# Import libraries

import pandas as pd
import numpy as np
import os

### Step 3. Import the data.

path = '/Users/vinhdao/Documents/Instacart Basket Analysis'

df_ords_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

### Step 4: Check the Output of Dataframes.

#Check Output: df_prods
df_prods.head()

df_prods.describe()

df_prods.shape

#Check Output: df_ords_combined
df_ords_combined.head()

df_ords_combined.shape

df_ords_combined.describe()

### Step 5. Merge df_prods with df_ords_combined

#Merge df_prods and df_ords_combined
df_merged_large = df_prods.merge(df_ords_combined, on = 'product_id')

df_merged_large.head(10)

df_merged_large['_merge'].value_counts()


### Step 7: Export files

df_merged_large.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))
