==> 03 Scripts/7. Deriving new Variables.ipynb <==
# CAREERFOUNDRY 4.7

# Import libraries
import pandas as pd
import numpy as np
import os

path = '/Users/vinhdao/Documents/Instacart Basket Analysis'

df_large = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))

# Create a subset of the first 1,000,000 rows
df = df_large[:1000000]

# Display the first few rows of the subset to verify
df.head()

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range product'
  else: return 'Not enough data'

# Apply the function

df['price_range'] = df.apply(price_label, axis = 1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna=False)

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.head(4)

df_large.loc[df_large['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_large.loc[(df_large['prices'] > 5) & (df_large['prices'] <= 15), 'prices_range_loc'] = 'Mid-range product'

df_large.loc[df_large['prices'] <= 5, 'prices_range_loc'] = 'Low-range product'

df_large['price_range_loc'].value_counts(dropna = False)

df_large[:15]

For this exercise, I dont know why there are so much Nan value

## Create the busiest day

# Change "order_dow" column name to "order_day_of_week."

df_large.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# Find frequency of orders_day_of_week.

df_large['orders_day_of_week'].value_counts(dropna = False)

# Use for-loop to determine whether orders are on "busiest day" (0 = Saturday), "least busy" (4 = Wednesday), or "regularly busy" (other days of the week).

result = []

for value in df_large["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

# Create new column in df_large to view results in context.

df_large['busiest_day'] = result

df_large['busiest_day'].value_counts(dropna = False)

### Step 2: Update busiest_day

# Create column “busiest_days" from "orders_day_of_week" to identify two busiest days (0, 1), slowest days (4, 3), or regularly busy (all other days).

result_modified = []

for value in df_large["orders_day_of_week"]:
  if value == 0 or value == 1:
    result_modified.append("Busiest days")
  elif value == 4 or value == 3:
   result_modified.append("Slowest days")
  else:
   result_modified.append("Regularly busy")

result_modified

### Step 3: Check the column value for accuracy

df_large['busiest_day'] = result_modified

df_large['busiest_day'].describe()

df_large['busiest_day'].value_counts(dropna = False)

df_large.head()

From the table, we can see that the busiest day grouping is match with the value of the number of orders_day_of_week

### Step 4:

df_large['order_hour_of_day'].value_counts()

# Use for-loop to find the frequency of the most hour of the day

result_hour = []

for value in df_large["order_hour_of_day"]:
  if value in [10, 11, 14, 15, 13, 12, 16, 9]:
    result_hour.append("Most orders")
  elif value in [23, 6, 0, 1, 5, 2, 4, 3]:
    result_hour.append("Fewest orders")
  else:
    result_hour.append("Average orders")

result_hour[:10]

df_large['busiest_period_of_day'] = result_hour

### Step 5:

# Print the frequency

df_large['busiest_period_of_day'].value_counts(dropna = False)

df_large.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_derived.pkl'))
