==> 03 Scripts/8. Grouping Data & Aggregating Variables.ipynb <==
# Step 1

# Import libraries

import pandas as pd
import numpy as np
import os

path = '/Users/vinhdao/Documents/Instacart Basket Analysis'
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_derived.pkl'))

# Step 2

# creat subset
ords_prods_merge_sub=ords_prods_merge[:1000000]

ords_prods_merge_sub.groupby(['department_id']).agg({'order_number':['mean']})

# Use groupby and agg functions to calculate mean of user_order_number
ords_prods_merge.groupby(['department_id']).agg({'order_number':['mean']})

## Step 3:

There are 21 distinct department IDs in the full dataset, while only 8 are present in the subset of 100,000 entries. Additionally, the average number of orders varies across departments when considering the entire dataset

## Step 4:

# Creating new column "max_order"

ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

# Checking for new column
ords_prods_merge.head(20)

# Assign loyalty flag
ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# Check loyalty flags
ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

ords_prods_merge[['user_id', 'loyalty_flag', 'order_number','max_order']].head(20)

## Step 5

# the min max and mean of each loyalty flag
ords_prods_merge.groupby(['loyalty_flag']).agg({'prices':['mean','min','max']})

New customers have the lowest product prices, both in terms of average and maximum. Loyal customers, on the other hand, have the highest product prices in both average and maximum values.

## Step 6:

# Creating column for average spender
ords_prods_merge['average_spend']=ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

# Creating flag for low and high spender
ords_prods_merge.loc[ords_prods_merge['average_spend']<10,'spender_flag']='Low_spender'

ords_prods_merge.loc[ords_prods_merge['average_spend']>= 10, 'spender_flag'] = 'High_spender'

# Check spender flags
ords_prods_merge['spender_flag'].value_counts(dropna = False)

pd.options.display.max_rows = 100

ords_prods_merge[['user_id', 'average_spend', 'spender_flag', 'prices']].head(100)

## Step 7:

# Create a column for frequet users
ords_prods_merge['Customer_frequency'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

# Creating flag on order frequency
ords_prods_merge.loc[ords_prods_merge['Customer_frequency'] > 20, 'frequency_flag'] = 'Non_frequent_customer'

ords_prods_merge.loc[(ords_prods_merge['Customer_frequency']>10) & (ords_prods_merge['Customer_frequency']<=20),'frequency_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['Customer_frequency'] <= 10, 'frequency_flag'] = 'Frequent Customer'

# Check frequency flags

ords_prods_merge['frequency_flag'].value_counts(dropna = False)

ords_prods_merge

ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'aggregated_ords_prods_merge.pkl'))
