==> 03 Scripts/9.1 Checking before Visualizing.ipynb <==
# Part 1:

# Import data libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

path = '/Users/vinhdao/Documents/Instacart Basket Analysis'

# Import new customer data

customers = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

## Step 3: Checking the dataset

# Checking the data set

customers.head()

customers.shape

customers.describe()

customers.info()

## Step 4: Wrangling the dataset

# Rename First Name as first_name

customers.rename(columns = {'First Name' : 'First_Name'}, inplace = True)

# Rename Surnam as surname

customers.rename(columns = {'Surnam' : 'Surname'}, inplace = True)

# Change the capital of STATE become State

customers.rename(columns = {'STATE' : 'State'}, inplace = True)

# Rename fam_status as marital_status

customers.rename(columns = {'fam_status' : 'marital_status'}, inplace = True)

## Step 5: Checking data quality and consistency

# Descriptive analysis

customers.describe()

# Checking the missing data

customers.isnull().sum()

As the result suggested, there are 11259 value missing the first name of the customers. This is not the key values, so we don't need to address more. Howevers, for further learning for this, we will go further deeper for the understanding

df_nan = customers[customers['first_name'].isnull() == True]

df_nan

Because these are the name of customers, we use any method to addressing the missing values. For this, we will leave it 

# Checking duplicates

df_dups = customers[customers.duplicated()]

df_dups

As the result suggested, there are no duplicated in the dataset

df_test = pd.DataFrame()

for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# Let's check again the data type
customers.dtypes

From above, we can see that all the data types is consistent and no need to change

## Step 6: Combine the dataset

# Import the rest of prepared Instacart data

ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'aggregated_ords_prods_merge.pkl'))

ords_prods_merge.head(10)

ords_prods_merge.dtypes

ords_prods_merge = ords_prods_merge.drop(columns = ['_merge'])

# Merge the dataset

ords_prods_customers = ords_prods_merge.merge(customers, on = 'user_id', indicator = True)

# Export the dataset

ords_prods_customers.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'aggregated_ords_prods_merge.pkl'))
